Scripts_to_be_checked/requirements.txt
awkward0==0.15.5
cachetools==4.2.4
cycler==0.11.0
kiwisolver==1.3.1
matplotlib==3.3.4
nested-dict==1.61
numpy==1.19.5
pandas==1.1.5
Pillow==8.4.0
pyinotify==0.9.6
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
pyzmq==23.2.0
scipy==1.5.4
six==1.16.0
uproot3==3.14.4
uproot3-methods==0.10.1
zmq==0.0.0
